+++ alicloud/provider.go
@@ -142,6 +142,7 @@ func Provider() terraform.ResourceProvider {
 			"alicloud_ots_table":                   resourceAlicloudOtsTable(),
 			"alicloud_cms_alarm":                   resourceAlicloudCmsAlarm(),
 			"alicloud_log_project":                 resourceAlicloudLogProject(),
+			"alicloud_log_store":                   resourceAlicloudLogStore(),
 		},
 
 		ConfigureFunc: providerConfigure,